Improved Components for ConfigProposalForm
